requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
flake8==3.8.3
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
mccabe==0.6.1
oauthlib==3.1.0
pathspec==0.8.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.5
PyYAML==5.3.1
regex==2020.6.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.26.2
yamllint==1.23.0
zipp==3.1.0
